Deleted the big "game" scene. Levels can be played by opening them up individually.
Also made an area2d that right now has the godot icon that acts as a next level thingy. It has code that when touched it is close to have the potential to open the next level, given that the levels are named as an integer. This is good for linear levels, but a more direct transition should be done for our purposes in the future.